test: :test_tube: delete old test files
