+++ tests/utils/helpers.ts
@@ -1,5 +1,8 @@
 import { ethers } from "hardhat";
 
+export const abiCoder = ethers.utils.defaultAbiCoder;
+export const provider = ethers.provider;
+
 export const EMPTY_PAYLOAD = "0x";
 export const DUMMY_PAYLOAD = "0xaabbccdd123456780000000000";
 export const ONE_ETH = ethers.utils.parseEther("1");
